requirements.txt
django==5.1
pillow==10.4.0
celery==5.4.0
pytest-django==4.8.0
pytest~=8.3.2
djangorestframework~=3.15.2
factory_boy==3.3.1
coverage==7.6.1
drf-spectacular==0.27.2
social-auth-app-django==5.4.2
django-jet-reboot==1.3.7
django-imagekit==5.0.0
sentry-sdk==2.13.0
redis-server==6.0.9
django-redis==5.4.0
django-cacheops==7.0.2